Rename commonsysfs -> genericlinux
