+++ components/board/genericlinux/board_test.go
@@ -1,4 +1,4 @@
-package commonsysfs
+package genericlinux
 
 import (
 	"context"
@@ -19,7 +19,7 @@ func TestRegisterBoard(t *testing.T) {
 	RegisterBoard("test", map[int]GPIOBoardMapping{}, false)
 }
 
-func TestCommonSysFs(t *testing.T) {
+func TestGenericLinux(t *testing.T) {
 	ctx := context.Background()
 
 	gp1 := &gpioPin{b: &sysfsBoard{
@@ -107,7 +107,7 @@ func TestCommonSysFs(t *testing.T) {
 		test.That(t, gn1, test.ShouldBeNil)
 	})
 
-	t.Run("test commonsysfs gpio pin functionality", func(t *testing.T) {
+	t.Run("test genericlinux gpio pin functionality", func(t *testing.T) {
 		err := gp2.SetPWM(ctx, 50, nil)
 		test.That(t, err, test.ShouldBeNil)
