+++ components/board/jetson/board.go
@@ -6,7 +6,7 @@ import (
 	"github.com/pkg/errors"
 	"periph.io/x/host/v3"
 
-	"go.viam.com/rdk/components/board/commonsysfs"
+	"go.viam.com/rdk/components/board/genericlinux"
 )
 
 const modelName = "jetson"
@@ -16,11 +16,11 @@ func init() {
 		golog.Global().Debugw("error initializing host", "error", err)
 	}
 
-	gpioMappings, err := commonsysfs.GetGPIOBoardMappings(modelName, boardInfoMappings)
-	var noBoardErr commonsysfs.NoBoardFoundError
+	gpioMappings, err := genericlinux.GetGPIOBoardMappings(modelName, boardInfoMappings)
+	var noBoardErr genericlinux.NoBoardFoundError
 	if errors.As(err, &noBoardErr) {
 		golog.Global().Debugw("error getting jetson GPIO board mapping", "error", err)
 	}
 
-	commonsysfs.RegisterBoard(modelName, gpioMappings, true)
+	genericlinux.RegisterBoard(modelName, gpioMappings, true)
 }
