+++ components/board/pi/impl/board.go
@@ -23,7 +23,7 @@ import (
 	"go.viam.com/utils"
 
 	"go.viam.com/rdk/components/board"
-	"go.viam.com/rdk/components/board/commonsysfs"
+	"go.viam.com/rdk/components/board/genericlinux"
 	picommon "go.viam.com/rdk/components/board/pi/common"
 	"go.viam.com/rdk/components/generic"
 	"go.viam.com/rdk/config"
@@ -42,7 +42,7 @@ func init() {
 			config config.Component,
 			logger golog.Logger,
 		) (interface{}, error) {
-			boardConfig, ok := config.ConvertedAttributes.(*commonsysfs.Config)
+			boardConfig, ok := config.ConvertedAttributes.(*genericlinux.Config)
 			if !ok {
 				return nil, rdkutils.NewUnexpectedTypeError(boardConfig, config.ConvertedAttributes)
 			}
@@ -55,7 +55,7 @@ func init() {
 type piPigpio struct {
 	generic.Unimplemented
 	mu            sync.Mutex
-	cfg           *commonsysfs.Config
+	cfg           *genericlinux.Config
 	duty          int // added for mutex
 	gpioConfigSet map[int]bool
 	analogs       map[string]board.AnalogReader
@@ -74,7 +74,7 @@ var (
 )
 
 // NewPigpio makes a new pigpio based Board using the given config.
-func NewPigpio(ctx context.Context, cfg *commonsysfs.Config, logger golog.Logger) (board.LocalBoard, error) {
+func NewPigpio(ctx context.Context, cfg *genericlinux.Config, logger golog.Logger) (board.LocalBoard, error) {
 	// this is so we can run it inside a daemon
 	internals := C.gpioCfgGetInternals()
 	internals |= C.PI_CFG_NOSIGHANDLER
